% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.readpaths-package.R
\docType{package}
\name{party.readpaths}
\alias{party.readpaths}
\alias{party.readpaths-package}
\title{party.readpaths: Package to add a readTerminalNodePaths function to party}
\description{
the readTerminalNodePaths function provides a data.frame with terminal node number and it's corresponding path.
}
