% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_paths.R
\name{readCtreePaths}
\alias{readCtreePaths}
\title{extract the terminal node paths from a conditional inference tree}
\usage{
readCtreePaths(ct, dts)
}
\arguments{
\item{ct}{A ctree.}

\item{dts}{A data.frame with the data that was the base for the ctree}
}
\value{
A data.frame with number of the terminal node and it's path
}
\description{
extract the terminal node paths from a conditional inference tree
}
\examples{
airq <- subset(airquality, !is.na(Ozone))
act <- party::ctree(Ozone ~ .,data = airq)
readCtreePaths(act, airq)
}
